pyproject.toml
[tool.poetry]
name = "anti-ddos"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "anti_ddos"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.1"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
